refactor(cketh): replace GetLogsParam with GetLogsArgs
