+++ rs/ethereum/cketh/minter/src/eth_rpc.rs
@@ -264,23 +264,6 @@ impl std::str::FromStr for BlockSpec {
     }
 }
 
-/// Parameters of the [`eth_getLogs`](https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_getlogs) call.
-#[derive(Clone, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GetLogsParam {
-    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
-    pub from_block: BlockSpec,
-    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
-    pub to_block: BlockSpec,
-    /// Contract address or a list of addresses from which logs should originate.
-    pub address: Vec<Address>,
-    /// Array of 32 Bytes DATA topics.
-    /// Topics are order-dependent.
-    /// Each topic can also be an array of DATA with "or" options.
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub topics: Vec<Topic>,
-}
-
 /// A topic is either a 32 Bytes DATA, or an array of 32 Bytes DATA with "or" options.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize)]
 #[serde(untagged)]
